Refactor OpenAI proxy code to include support for Google Search and simplify model mapping
