+++ src/types.ts
@@ -17,7 +17,7 @@ type OPEN_AI_MODELS =
   | 'gpt-3.5-turbo-0613'
   | 'gpt-3.5-turbo-16k-0613';
 
-type IProxy = (action: string, body: any, env: Env) => Promise<Response>;
+type IProxy = (action: string, body: any, env: Env, builtIn?: boolean) => Promise<Response>;
 
 interface AzureKey {
   resourceName: string;
